requirements.txt
# https://docs.langflow.org/get-started-installation#datastax-langflow
langflow
python-dotenv